Snapshot of cl/561165356
