+++ clif/testing/python/CMakeLists.txt
@@ -233,6 +233,6 @@ add_pyclif_library_for_test(property_get_set_cpp_derived property_get_set_cpp_de
 
 add_pyclif_library_for_test(python_multiple_inheritance python_multiple_inheritance.clif)
 
-add_pyclif_library_for_test(mock_uninitializable mock_uninitializable.clif
-  PY_DEPS clif_python_instrumentation_for_testing
-)
+add_pyclif_library_for_test(mock_uninitializable mock_uninitializable.clif)
+
+# OMITTED: instrumentation_for_testing_test (pytype currently not supported on GitHub)
